# apps/backend/apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tenants": {
      "get": {
        "description": "전체 테넌트 목록을 조회합니다.",
        "operationId": "TenantsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "테넌트 목록 조회",
        "tags": [
          "테넌트"
        ]
      }
    },
    "/tenants/{id}": {
      "get": {
        "description": "ID로 단일 테넌트를 조회합니다.",
        "operationId": "TenantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "테넌트 상세 조회",
        "tags": [
          "테넌트"
        ]
      },
      "patch": {
        "description": "테넌트 정보를 수정합니다.",
        "operationId": "TenantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "테넌트 수정",
        "tags": [
          "테넌트"
        ]
      },
      "delete": {
        "description": "테넌트를 삭제합니다(소프트 삭제).",
        "operationId": "TenantsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "테넌트 삭제",
        "tags": [
          "테넌트"
        ]
      }
    },
...
  },
  "info": {
    "title": "blommunity API 문서",
    "description": "blommunity의 REST API 문서입니다. 인증, 게시물 관리 등의 기능을 제공합니다.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TenantEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "테넌트 ID"
          },
          "name": {
            "type": "string",
            "description": "테넌트 이름"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "생성일시"
          }
        },
        "description": "테넌트 조회 응답에 포함되는 테넌트 정보",
        "required": [
          "id",
          "name",
          "createdAt"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "테넌트 이름"
          }
        },
        "description": "테넌트 수정 요청 시 전달하는 정보 (모든 필드 선택)"
      },
...
    }
  }
}